Library cleanup.
It's time, I think, to revist this library and make it something to be proud of.
I made some curious decisions in my previos work--largely to explore using
proper--which I would not now make.

This commit does some simple cleanup, removing old, busted tests and leaves the
field open. The next bit of work will be to make this more sound.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>