Introduce benchmarking.
The bit of this commit that bothers me is the introduction of end_of_string
handling in internal partial_matches/3. I'm not sure how it could happen that
the zipper is forwarded but the effectively randomized testing of
benchmark_test_() indicates that it does.

It'd be good to comment out the benchmark tests, take the coverage ding and
figure out how to reproduce in a specific test for that function.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>